requirements.txt
-i https://pypi.org/simple
pymqi==1.12.10
python-dotenv==1.0.1; python_version >= '3.8'
